completed revision trees.

+++ BST/pratice1.js
@@ -26,12 +26,15 @@ class BinarySearchTree {
     }
   }
 
-  dlelete(key){
-      this.root = this.deleteNode(this.node,key);
+  delelete(key){
+      this.root = this.deleteNode(this.root,key);
   }
 
   deleteNode(node,key){
-    if(key<node.key){
+    if(!node){
+      return null;
+    }
+    if(key<node.key){ 
       node.left= this.deleteNode(node.left,key);
     }else if(key>node.key){
       node.right  = this.deleteNode(node.right,key);
@@ -43,9 +46,9 @@ class BinarySearchTree {
       }else if(node.right===null){
         return node.left;
       }else{
-        let tempNode = this.findMaxNode(this.left);//*implimenting pre order predesussser.
+        let tempNode = this.findMaxNode(node.left);//*implimenting pre order predesussser.
           node.key = tempNode.key;
-          node.right = this.deleteNode(node.right,tempNode.key);
+          node.left = this.deleteNode(node.left,tempNode.key);
       }
     }
     return node;
@@ -84,3 +87,20 @@ BinarySearchTree.prototype.insert = function (key) {
     this.insertNode(this.root, newNode);
   }
 };
+
+let newTree = new BinarySearchTree();
+
+newTree.insert(23);
+newTree.insert(34);
+newTree.insert(12);
+newTree.insert(89);
+newTree.insert(43);
+newTree.insert(8);
+newTree.insert(53);
+newTree.insert(7);
+newTree.insert(3);
+newTree.insert(63);
+
+newTree.delelete(23);
+console.log(newTree.root);
+console.log(newTree.preOrederTraversal());
